requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.0
pillow>=9.3.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
pandas>=1.5.0
facenet-pytorch>=2.5.2
opencv-python>=4.6.0
tqdm>=4.64.0
albumentations>=1.3.0
optuna>=3.0.0
kagglehub>=0.1.5
mtcnn==0.1.1
seaborn>=0.12.0
psutil>=5.9.0 